final_pjt_back/requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.57.0
django-cors-headers==4.3.1
django-crontab==0.7.1
django-environ==0.11.2
djangorestframework==3.14.0
exceptiongroup==1.1.3
h11==0.14.0
HTMLParser==0.0.2
idna==3.4
joblib==1.3.2
numpy==1.26.2
oauthlib==3.2.2
outcome==1.3.0.post0
pandas==2.1.3
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
threadpoolctl==3.2.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wsproto==1.2.0
